Restructure so tests pass
