+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 # replace with your username:
-name = python-calculator
+name = pythoncalculator
 version = 0.0.1
 author = Example Author
 author_email = author@example.com
@@ -16,10 +16,5 @@ classifiers =
     Operating System :: OS Independent
 
 [options]
-package_dir =
-    = src
 packages = find:
 python_requires = >=3.6
-
-[options.packages.find]
-where = src
