+++ tests/test_add.py
@@ -1,5 +1,5 @@
-import src.python_calculator.add as add
+from pythoncalculator import add
 
 
 def test_add():
-    assert add.add(1, 3) == 4
+    assert add(1, 3) == 4
